# Cache Windows mount point info