@harshavardhana 2019-11-25
It's not a regression as such @klauspost it is just this is expected on Windows if hit the disk to get this, I have seen that performance is dismal compared to `statvfs` call on Linux.